Feat/interaction improvements (#2)
* User Warnings (Initial)

The program now warns the user that there aren't any audios to set photos to.  I've also added an Errors class (that just shows the error to the user) and an enum that will get filled with "user visible" errors.

* Audio/Photo Split

Split up the audio and photo selector buttons. The backend still calls the same code, since I went to the trouble of letting that function handle both photos and audio. May split up the backend after I've gotten to another release iteration.

* User Warnings (Additional)

Just a few extra warning stages shown to the user. I'm thinking the next steps here are sizing the actual soundboard to the height/width of the buttons, and to add support for multiple soundboard properties files, but that second idea needs to happen in another branch.

* Dynamic Display Improvement

I've made it so that the dynamic content windows respond to the size of everything placed in it, so now their display isn't astronomically large/far too small.

* Constants/Titles

Changed some numbers to constants, and added titles to all the windows. I also removed a whole class that wasn't getting used, as it was replaced in the process of splitting the adding of photos and audios into two different buttons