Functionality
After a great deal of fiddling and finagling, I've managed to get the program to do what I intended it to do. In short, it runs through my directory where I keep properties files, and lets the user pick from the soundboards that live there. Additionally, I spotted a small bug in the method I use to arrange buttons on the board, but it was a simple fix, so I got it out of the way now. Barring additional bugs cropping up in testing, this is ready to merge into develop.