Bug Fix
Alas, I should have knocked on wood saying I was ready for a merge, because I found a second bug in the way I handle the sorting of buttons. Squashed now! Going to test some more just in case there's something else that's getting out of my view, but for real this time, if I don't find anything, it's merge ready.