Feat/multiple soundboards (#3)
* Choosing a Soundboard (untested)

I usually wouldn't push untested code, but it's time for a lunch break. What it's intended to do is, when the user asks to choose a soundboard, it sets the filename we pull all the information from to the user's choice.

* code for your hands

* Tinkering

For some reason (I'm still pinning down why) choosing an existing soundboard properties file doesn't show an existing board. Additionally, I'm having trouble writing properties files to disk. Problem for tomorrow's me, I suppose, I'm clocking out.

* Functionality

After a great deal of fiddling and finagling, I've managed to get the program to do what I intended it to do. In short, it runs through my directory where I keep properties files, and lets the user pick from the soundboards that live there. Additionally, I spotted a small bug in the method I use to arrange buttons on the board, but it was a simple fix, so I got it out of the way now. Barring additional bugs cropping up in testing, this is ready to merge into develop.

* Bug Fix

Alas, I should have knocked on wood saying I was ready for a merge, because I found a second bug in the way I handle the sorting of buttons. Squashed now! Going to test some more just in case there's something else that's getting out of my view, but for real this time, if I don't find anything, it's merge ready.

* Assorted Wrinkle Ironing

I gave everything a thorough top-to-bottom test, and found a few small things to fix. Up next in features to be built: directly opening a soundboard (so you don't have to select every time) and deleting soundboards (deleting specific files might be a bit much, we'll see)