Deletion of individual items
You can now delete individual items! There was also a change in the showing of sounds, because I forgot to update the "empty of keys" number to be 1 (since the properties file will always have a key stating that its for a soundboard). Otherwise, works as intended, and catches the user error of not having any sounds from an earlier feat branch. Ready for bug test/merge to dev, which then should be ready for release! If I've got time, I'll do that tomorrow.