Feat/delete button (#5)
* Delete Boards

You can delete whole soundboards! Gosh golly! This was a bit of a pain, since I had to go use a different method of deletion than the one that's built into the normal File import. However, the real pain is yet to come: deleting single sounds.

* Deletion of individual items

You can now delete individual items! There was also a change in the showing of sounds, because I forgot to update the "empty of keys" number to be 1 (since the properties file will always have a key stating that its for a soundboard). Otherwise, works as intended, and catches the user error of not having any sounds from an earlier feat branch. Ready for bug test/merge to dev, which then should be ready for release! If I've got time, I'll do that tomorrow.

* Initial Testing

Some small bugs/user error catching found these bugs to squash, but things still need a full top-to bottom test. It's lunchtime, though, so time to commit and come back to it later.

* Bug Fixes/Updated HowTo

Ironed out the last of the bugs that I know of, and changed the how to, so that it actually describes how to do all the things. I'll be updating the readme with that howto when this reaches master branch.