Release 1.0.0 (#6)
* Feat/visual changes (#1)

* Organization

A small aesthetic change that organises the buttons into rows.

* Better Photo Choosing

The original method of selecting a photo for a file was dreadfully unweildy, so I added something a bit more user-friendly. This is only pretty loosely a visual change, so it's time for a new branch.

* Feat/interaction improvements (#2)

* User Warnings (Initial)

The program now warns the user that there aren't any audios to set photos to.  I've also added an Errors class (that just shows the error to the user) and an enum that will get filled with "user visible" errors.

* Audio/Photo Split

Split up the audio and photo selector buttons. The backend still calls the same code, since I went to the trouble of letting that function handle both photos and audio. May split up the backend after I've gotten to another release iteration.

* User Warnings (Additional)

Just a few extra warning stages shown to the user. I'm thinking the next steps here are sizing the actual soundboard to the height/width of the buttons, and to add support for multiple soundboard properties files, but that second idea needs to happen in another branch.

* Dynamic Display Improvement

I've made it so that the dynamic content windows respond to the size of everything placed in it, so now their display isn't astronomically large/far too small.

* Constants/Titles

Changed some numbers to constants, and added titles to all the windows. I also removed a whole class that wasn't getting used, as it was replaced in the process of splitting the adding of photos and audios into two different buttons

* Feat/multiple soundboards (#3)

* Choosing a Soundboard (untested)

I usually wouldn't push untested code, but it's time for a lunch break. What it's intended to do is, when the user asks to choose a soundboard, it sets the filename we pull all the information from to the user's choice.

* code for your hands

* Tinkering

For some reason (I'm still pinning down why) choosing an existing soundboard properties file doesn't show an existing board. Additionally, I'm having trouble writing properties files to disk. Problem for tomorrow's me, I suppose, I'm clocking out.

* Functionality

After a great deal of fiddling and finagling, I've managed to get the program to do what I intended it to do. In short, it runs through my directory where I keep properties files, and lets the user pick from the soundboards that live there. Additionally, I spotted a small bug in the method I use to arrange buttons on the board, but it was a simple fix, so I got it out of the way now. Barring additional bugs cropping up in testing, this is ready to merge into develop.

* Bug Fix

Alas, I should have knocked on wood saying I was ready for a merge, because I found a second bug in the way I handle the sorting of buttons. Squashed now! Going to test some more just in case there's something else that's getting out of my view, but for real this time, if I don't find anything, it's merge ready.

* Assorted Wrinkle Ironing

I gave everything a thorough top-to-bottom test, and found a few small things to fix. Up next in features to be built: directly opening a soundboard (so you don't have to select every time) and deleting soundboards (deleting specific files might be a bit much, we'll see)

* Full Change (#4)

This was a bit of a small change for a full branch, but I didn't want to just lump it into the deleting things branch, since that's bound to be a LOT of change.

* Feat/delete button (#5)

* Delete Boards

You can delete whole soundboards! Gosh golly! This was a bit of a pain, since I had to go use a different method of deletion than the one that's built into the normal File import. However, the real pain is yet to come: deleting single sounds.

* Deletion of individual items

You can now delete individual items! There was also a change in the showing of sounds, because I forgot to update the "empty of keys" number to be 1 (since the properties file will always have a key stating that its for a soundboard). Otherwise, works as intended, and catches the user error of not having any sounds from an earlier feat branch. Ready for bug test/merge to dev, which then should be ready for release! If I've got time, I'll do that tomorrow.

* Initial Testing

Some small bugs/user error catching found these bugs to squash, but things still need a full top-to bottom test. It's lunchtime, though, so time to commit and come back to it later.

* Bug Fixes/Updated HowTo

Ironed out the last of the bugs that I know of, and changed the how to, so that it actually describes how to do all the things. I'll be updating the readme with that howto when this reaches master branch.

* Updated ReadMe

Updated the readme with the new how-to's text. Provided that I can get the program to run as intended after packaging it into a jar, this is ready for merge into master (i.e. release).